Basic layout of new window
